[SymForce] Cache factor indexes in the Linearizer
Instead of in the factors.  This means we no longer need a notion of a
Values Id either, since a Linearizer is only valid for a single
structure of a Values.

This also means if you are linearizing a Factor _outside_ an optimizer,
you'll need to create and persist this index entry cache yourself, if
you don't want to do repeated unordered_map lookups.

This is a bit suboptimal, in that if you want to linearize a factor you
must either have this cached, or you'll allocate a vector to store it.
Avoiding this isn't easy though, we'd have to pipe two paths through the
factor creation methods in `factor.tcc`, so that you can either pass
through a vector of `index_entry_t` that's cached and passed in at
linearization time, or access the vector of keys stored in the factor
itself.  And you'd have to specify this option in the polymorphic
function signatures `Factor::HessianFunc` and `Factor::JacobianFunc`,
which would make those super gross.  I think this is a reasonable
tradeoff.

It might be mildly better if those functions could take a
`span<index_entry_t>`, so that you could use something that's not
heap-allocated if you wanted to.

[Testing] Ran the symforce_optimizer_test under callgrind, verified
`sym::Values::IndexEntryAt(const sym::Key&)` is not called from the
optimization loop

Fixes symforce-org/symforce#266

Reviewers: bradley,nathan,chao,hayk
Topic: sym-factor-cache
GitOrigin-RevId: 795efacbc57717b30f9dcde8d557b3fa0b1adce7