[SymForce] CUDA Backend
Add a CUDA backend to SymForce.  This currently only supports scalar,
vector, matrix, and databuffer types, not geo and cam.

Inputs and outputs are passed by pointer, or optionally as CUDA vector
types.  Matrices are assumed to be flattened row-major, which is maybe a
little unintuitive. Pointer arguments don't support strides, but we can
probably add that when we need it?

Functions can be generated either as inline, or with a separate `.cu`
file containing the definition.

[Testing] Added codegen tests

Fixes symforce-org/symforce#79

Reviewers: bradley,nathan,chao,hayk,peter,ryan,dancernea
Topic: sym-cuda
GitOrigin-RevId: fa5dbd13b450770af2ea6e457cd7d6dde4c53f36