[symforce] Add geo.Unit3 to model directions in R^3
Direction in R^3, represented as a Rot3 that transforms [0, 0, 1] to the desired direction.

Note: an alternative implementation could direct store a unit vector and define its boxplus
manifold as described in Appendix B.2 of (Hertzberg 2013). This is essentially done by finding
the Householder reflector of x and use it to transform a small delta. Namely,

x.retract(delta) = x [+] delta = Rx * Exp(delta), where
Exp(delta) = [sinc(||delta||) * delta, cos(||delta||)]

However, this is problematic in that it has a singularity at [0, 0, 1] and this singularity,
albeit can be removed in the retract function, still shows up in the `storage_D_tangent`
function. Unless we change the signature of `storage_D_tangent`, this will be problematic for
symforce. Therefore, we make a compromise here by using a rotation to represent a direction in
R^3, such that the unit vector corresponding to this direction can be recovered by transforming
e_z by this rotation.

Revisions:
 - Some workaround to successfully generate python and c++ code
 - Rename Unit3R -> Unit3
 - Add a Random function
 - Add custom generated function `from_unit_vector`
 - Add more tests
 - Make Unit3.from_unit_vector() more numerically stable by normalizing the input
 - Add full reference on Hertzberg 2013
 - Flip tangent space xy such that retraction at identity matches unit vector representation
 - Fix symforce_values_test tolerance
 - Fix linter issues
 - Add missing include in FUNCTION.h.jinja
 - Fix various typos
 - Add a `to_rotation()` method

Topic: chaoqu_symforce_unit3r
Reviewers: aaron-skydio, hayk-skydio, bradley-solliday-skydio
GitOrigin-RevId: 0368bb48f58d39fa6677a743d4794c675bf52156