[SymForce] Better PyTorch test coverage
Use the tools from the CUDA backend (in particular
backend_test_function) to test the PyTorch backend.  Caught some issues
that I fixed here as well :)

Added a test that also calls pytorch-generated functions, it only runs
if PyTorch is installed (same as we do for numba), and I'm not adding
pytorch as a symforce dependency so it won't run in symforce CI.  I
think this is fine.

Reviewers: bradley,nathan,chao,hayk,john-m
Topic: sf-pytorch-tests
Relative: sym-cuda
GitOrigin-RevId: ecd84448ccc1358f8063807d7c91cdfa59eb13d3