[SymForce-External] [ImuFactor] Add an on manifold Imu Factor
The core use-facing api is found in
- symforce/slam/imu_preintegration/preintegrated_imu_measurements.h
- symforce/slam/imu_preintegration/imu_factor.h

Unlike most other factors (/residual functions) in symforce, the
ImuFactor is a functor so as to store the pre-integrated values
with the function rather than in the `Values`. (There are a lot of
parameters that are constant and unique to each factor).

[ImuFactor] Unit test the on manifold ImuFactor

Adds four tests meant to jointly test the
`ImuManifoldPreintegrationUpdate` function.

The calculated covariance is tested by sampling the values it is a
covariance of and comparing it to the numircally calculated covariance
from the samples. Warning, there is a tradeoff between time (number of
samples) and accuracy for the calculated covariance. To speed things up,
the sampling process is multi-threaded. Note, with my current parameters,
it takes about 2^23 samples to be pretty robust to changes in the random
seed. But that takes a minute to run on my desktop with 16 cores. So in
the interest of testing time, I've cut the sample size by a bit. Of
course, this means if you change the seed, the test might end up
failing. To know if it's really a problem, bump up the sample size and
see if it persists.

The derivatives wrt to bias are calculated by perturbing the input in
each direction to numerically calculate the columns of the derivative.

The actual change in state (DR, Dv, Dp) isn't tested in a very
intelligent manner. They didn't have any simple and obvious properties
that could be easily tested. Still, their implementations are pretty
simple, so I figured comparing to a C++ implementation would at the very
least protect against someone accidentally deleting a line in the
symbolic code or whatever.

Topic: on_manifold_imu_factor_test
Relative: on_manifold_imu_factor
Closes https://github.com/symforce-org/symforce/pull/279
Closes https://github.com/symforce-org/symforce/pull/280
GitOrigin-RevId: 31bcaf8dbcf7b967f03199fea71a66d99b5be52b