[SymForce-External] Set min version of lazy-object-proxy to 1.9.0
Set min version of lazy-object-proxy to 1.9.0

This is because lazy-object-proxy does not have built wheels on pypa for
linux on version 1.8.0, and for some reason running it's setup.py can
fail when done locally.

Also, for some reason, a pre-built wheel for the same package is not
available for python3.11 on bionic (even for version 1.9.0). Note,
lazy-object-proxy can be built on pip 22.3.1 (but not on the current
latest, 23.0.0). Suspect this is a bug on pip's end, but if not, then
it's probably a bug on lazy-object-proxy's end. Anyway, because we don't
want to contort our code around other package's problems, I am just
removing the python3.11 on bionic tests. If for some reason you do need
build on python3.11 on bionic, just set your pip version to 22.3.1 then
try again. Should work (not just for this pair but more generally).

Topic: get_CI_workflow_passing_again
Closes symforce-org/symforce#296
GitOrigin-RevId: 33a9d4a7475f0aea7e77886ef14a481a1aa64b21