[SymForce] Add `evict` option for `load_generated_function`
This lets you specify that you don't want to unload any modules, which
is necessary for numba to work.

I don't love this, open to discussing - I'd also like to get the
follow-up PRs that use this merged though (I might hack around this
there if we don't want to merge this to symforce).

Topic: sym-auto-numba
Reviewers: bradley,nathan,hayk,chao
GitOrigin-RevId: ddede36e200dbcf80d796791a7169965691de857