Use URL for FetchContent
Use URL and URL_HASH for everything, for the reasons described in the
comment.  Versions of everything are unchanged, except for tl::optional,
which is now on an actual release (1.1.0).

Tested locally that this works with no dependencies installed on the
system, with cmake 3.20 (cmake 3.19 doesn't seem to be available on pypi
anymore)

Topic: fetchcontent-url-download