[SymForce] Check types of geo constructor arguments
Previously if you passed 0.0 it would silently succeed, since 0 is
falsey.  And it would appear to do the right thing, typically, since
usually the intention is that you want the identity if you're passing 0
(e.g. constructing a Rot2 and expecting `Rot2.from_angle(0)`).  I think
we want to make you be explicit, it was accidental that this was
happening to do the right thing.  Other types already check this way

Reviewers: bradley,nathan,chao,hayk
Topic: geo-constructor-checks
GitOrigin-RevId: c6fbe2c8f7ba1d353976c10ff53cf11f190cf5ba