[SymForce] NumericFactor sets return_2d_vectors
This is required since it's perfectly valid for results of linearizing a
factor to have one of their dimensions be 1 (e.g. if the factor jacobian
is 1x6).  If we don't include both dimensions, pybind will convert to a
column vector, which means the jacobian is then the wrong shape, and you
get errors like in https://github.com/symforce-org/symforce/issues/308.

Fixes https://github.com/symforce-org/symforce/issues/308

Also added a test which this takes fail->pass, and a test of
robot_3d_localization as well

Reviewers: bradley,nathan,chao,hayk
Topic: fix-numeric-factor-2d
GitOrigin-RevId: 0d23cae5e2fba3095d048c26394c22de973c2f23