[SymForce] rm index from LinearizedFactor types
**Why**

First off, I simply thought the fact that we conditionally re-evaluate
the index when calling `Factor::Linearize` if the index is empty kind of
confusing, as it's behavior buried somewhat deeply in the call-graph and
it's not really something you would expect to be done just from the name
`Linearize`.

This ended up contorting the code around the call-site, leading us to
manually set the `index` of a linearized dense factor to `{}` without
much of an obvious explanation. --- It turned out the reason was to ensure
the index is re-calculated even when re-using a
`dense_linearized_factor_type` so that it can be used to calculate the
appropriate offsets to use when writing the linearized factor into the
problem linearization. If this isn't done, then these offsets will be
calculated incorrectly (if other runtime errors don't arise beforehand. (I
include this explanation to demonstrate how the extra-statefullness hides
important behavior, obscuring the code).

Second, the only place we really use the `dense_linearized_factor_t` is
in the `Linearizer`, which uses the same `linearized_dense_factor_t` for
multiple factors, each with  different incompatible indices. That is to
say, the field is invalid most of the time (the linearizer only works
because the parts that are different between factors are set to the
correct value right before it is needed).

While it works right now, it looks pretty easy to put the linearizer
into an invalid state if your not mindful of this.

**What I've done**

I've removed the index field from both the dense and sparse linearized
factor types.

Because some sanity checks on the dimensions of the factor linearization
are only possible with information from the index, I moved those checks
out of `Factor::Linearize` and into the call-site.

The index is now calculated by a free function and stored into a
temporary (and so no-longer lives for the lifetime of the optimization).

I also changed the formating of the linearized factor classes during
printing. This is actually arguably one downside of how I have things
now, as information that would plausibly be useful for debugging is now
no-longer bundled with the classes, and thus not present when printing.

Incidentally, I also changed it so that the sparse factors no longer
have their residuals stacked at the end of the problem residual, and
instead the order of sub-residuals in the problem match the order of
the factors as they were passed to the optimizer. This change isn't
essential, but made it easier to place the index of a sparse linearized
factor in a temporary that would only live for a short period of time.

Topic: no_index_in_linearized_factor
Relative: relinearize_linearization_reference
GitOrigin-RevId: 3f9ca65b7cfe2c47e4818785f9c8996d35af0b4c