[SymForce] Reduce # of index lookups in linearizer
**Where there was room for improvement**

We have been computing an index for the linearized keys of every single
factor. This was because we needed the `index_entry_t` for those keys in
order to know the tangent dimension of those keys to calculate their
offsets and sizes within the linearized factors.

However, most keys which are linearized by a factor are also being
optimized in the problem, and are therefore in the `state_index` (which
we have to read for each key in order to know their offsets within the
combined state vector). The `index_entry_t` in the `state_index` is the
exact same as the entry in the factor index, so in such cases there
really was no need to have created the index for the factor.

**What I did**

What I do in this commit is lookup a linearized key's index entry if and
only if it wasn't already in the `state_index`. Note, if it's not in the
`state_index`, we still need its tangent dim in order to calculate the
tangent dim of the whole factor (for sanity checks) and to calculate the
factor offsets of any subsequent optimized keys.

Because the per-factor index was no longer needed, I deleted
`GetIndexForKeys`.

In order to look up keys when they were not present in the
`state_index`, I added to `ComputeFactorHelper` the argument `Values
problem_values`. I passed in the whole `Values` rather than just the
`Values::MapType` because I thought a) it would be clearer what the
argument should be and b) `Values::IndexEntryAt` raises a helpful error
message when a key is not present (that I didn't want to raise
manually).

In order to know which keys are linearized by a factor and what order
they come in, I add the `linearized_keys` argument to
`ComputeFactorHelper`.

Because the factor tangent_dim is no longer being provided by the factor
index (and because I needed to compute it anyway for the offsets) I
changed the return value of `ComputeFactorHelper` to be a pair of the factor
tangent_dim and made the old return value (note, it bothers me that we
only partially fill out its fields. Suggests to me
that type should be split up). Also rearranged template parameters to
allow some of them to be deduced.

A side note, but we were performing a copy of the `factor_helper` for
each dense factor when we pushed it back (should have been moving it).
Fixed this too.

Topic: no_index_per_factor
GitOrigin-RevId: 5891922f0ae5f88aa3e8731db763e4b5192f4715