[SymForce] Add PreallocatedLinearizedDenseFactors
**Context**
Currently, if the same `LinearizedDenseFactor` is used as the output
argument to a Factor with different residual and rhs dimensions than
what it already has, Eigen will reallocate the wrong sized members, even
if the new size is smaller.

To avoid this, we ensure that we don't pass the same
`LinearizedDenseFactor` to `Factor`s with different sizes.

Simultaneously, to reduce the amount of memory used, we do not simply
have a seperate `LinearizedDenseFactor` for each dense `Factor` (as we
do for sparse factors), but instead have only one
`LinearizedDenseFactor` for each dense factor shape.

Before this commit, all of this logic was inlined in members of
`Linearizer` and in the body of `Linearizer::BuildInitialLinearization`.

**Motiviation**
The principal motivation for this commit is to allow the same logic to
be reused in a different linearizer class.

As a side benefit, I think it makes
`Linearizer::BuildInitialLinearization` easier to read.

**What I did**
I factored out the logic into `PreallocatedLinearizedDenseFactors`.

An instance of `PreallocatedLinearizedDenseFactors` is stored by the
linearizer, and contains the unique `LinearizedDenseFactors` as well as
the mapping from `Factor`s to the corresponding `LinearizedDenseFactor`.

Usage is detailed in doc-string (but also demoed in the linearizer.cc).

NOTE: requires the user pass in a very awkward temporary unordered_map
to the call to `AppendFactorSize`. I do it this way because I couldn't
think of a way to encapsulate this without either storing more data than
is necessary or adding a lot of extra boilerplate code.

Topic: extract_dense_linearized_factor_consolidation
GitOrigin-RevId: 69c672ec3228b13b4cc55a0a75f3b01dee74eccc