[SymForce] Correct init residual order
In a previous commit, I changed the order in which
`ComputeFactorHelper` is called on each factor to match the order of the
factor vector itself (previously, all of the dense factors were
processed, then all of the sparse factors).

This had the (known) effect of also changing the problem residual to
contain the factor residuals in the same order.

However, I forgot to ensure the initial residual (which is not
constructed using the output of `ComputeFactorHelper`) was also
reordered.

This commit fixes that by ensuring that the individual factors have
their residual added to the combined residual in the same order that
`ComputeFactorHelper` is called.

Also, I'm pretty sure the exact same issue is present for the jacobian,
which is processed by `UpdateFromSparseFactorIntoTripletLists`, so I
moved that as well.

Since the `rhs` doesn't depend on the order of factors within the
residual, it didn't need to be moved, but I moved it anyway to be more
consistent.

Topic: fix_initial_residual_order
Relative: factor_into_triplets_output_arg_by_reference
GitOrigin-RevId: 3ec08f462492baa225b21e4e7c6de60cbc6beeb5