[SymForce] Move ComputeKeysToOptimize to factor.h
I want to reuse this function for the dense linearizer, so I need it in
some neutral ground outside of `linearizer.h`.

As an aside, I think it makes more sense for this function to live in
`factor.h` regardless, so I'm moving it there.

Topic: factor_out_common_linearizer_code
GitOrigin-RevId: 7af7b60c5f4db2bc71160b5c2e5b09dc6f3b346c