[SymForce] Make optimization stats sparse agnostic
On it's own, this commit is not very well motivated, but it is in
preperation for adding a dense optimizer.

The plan is to change the `Linearization` type to allow dense jacobian
storage. In order to do that, the methods which assumed `jacobian` was a
sparse matrix had to be modified.

What I thought would be easiest (and what I did) was to remove the
methods and make them free functions. (Incidentally, one of them was
useful in another part of the Levenberg-Marquardt solver).

`JacobianColumnPointersMap` and `JacobianRowIndicesMap` were combined to
form `GetSparseStructure` because both are only ever used in that
structure.

`JacobianValuesMap` was changed to `ValuesOf`. The sparse
version returns a map to a vector, while the dense version just returns
a reference original jacobian (it works because they are both more or
less references, so they get used the same way. The implementation also
looks simpler this way).

The dense version of `GetSparseStructure` returns a default initialized
sparse structure because it doesn't ever make sense to read the sparse
structure of a dense matrix.

The intention is for the `jacobian_sparsity`, `linear_solver_ordering`,
and `cholesky_factor_sparsity` are intended to not be filled during a
dense solve, because these values become trivial in the dense case.

The `jacobian_values` field of `optimization_iteration_t` was changed to
a `MatrixXf` from a `VectorXf` because that allows the dense version to
store a whole jacobian with its shape (i.e., not flattened) which would
probably be more useful, while still allowing the sparse version to be
its flat CSC non-zero values.

Topic: sparse_agnostic_stats
GitOrigin-RevId: d1f66c214452e301aa55a2a28ba4b8ee54309f4e