[SymForce] Rename linearization_sparse_key_helper
Renaming `linearization_sparse_key_helper_t` to `linearization_offsets_t`.

I'd like to reuse this type for the dense linearizer, so I thought it
would be better to give it a name that is more suggestive of its
contents and is less specific about who is using it.

I'll be using it for more or less the exact same thing that the sparse
linearizer uses it for. (i.e., for figuring out where the components of
a factor's linearization get mapped to in the combined linearization).

Topic: rename_linearization_sparse_key_helper_t
Relative: fix_initial_residual_order
GitOrigin-RevId: adab6f10c6e3aaad66ca29de371db707a148a989