[SymForce] Template Linearization on sparsity
In preperation for a dense optimizer.

In order to do a dense linear solve, we need a dense matrix to solve.
For that reason want to be able to store dense matrices in our
`Linearization` type.

I thought the most sensible way to do this is to change the
`Linearization` type to be templated on the matrix type it stores.

Removed the `Scalar` template parameter because that can be deduced from
the Eigen matrix type.

In cases where the `Linearization` is always expected to be
sparse/dense, I added the typedefs `SparseLinearization` and
`DenseLinearization`.

Since the rest of the optimizer infrastructure is not yet set up to
allow dense optimization at the moment, I changed usages to use
`SparseLinearization`.

Topic: linearization_also_dense
Relative: rename_linearization_sparse_key_helper_t
GitOrigin-RevId: ea5a6569d476343c87a0204c80587f7f13b458ad