[SymForce] Add helpful err msg if unopt'zed factor
As demonstrated in https://github.com/symforce-org/symforce/issues/318
the AssertionError that's raised if the user attempts to create an
optimizer with a factor that doesn't touch any optimized keys isn't very
helpful.

Adds a test that the intersection of the factor keys and the keys to
optimize is not empty and raises a ValueError with extra information if
it is (the old `AssertionError` would have been raised otherwise).

Topic: err_msg_if_py_factor_unoptimized
GitOrigin-RevId: bed40079b586a82a2c3e7072cf7c5a8bdf91ceb3