[SymForce] Make LevenbergMarquardt Sparse agnostic
In order for the LevenbergMarquardt to work with both dense and sparse
linear solvers, it needs to be able to use both dense and sparse
linearizations.

To do this, I'm having LevenbergMarqaurdt query the linear solver for
what matrix type it should use to store the hessians (and thus what sort
of matrix the linear solver expects).

**Details**
- Had option between templating `LevenbergMarquardtState` on the
linearization type or the matrix type (of the linearization). I opted
for the matrix type just so it's more obvious what sort of type
`Linearization<MatrixType>` is than `LineatizationType`.

Topic: sparse_agnostic_opt
Relative: linearization_also_dense
GitOrigin-RevId: cd3bccd54092c747ef294cbe33fe723af757676c