[SymForce] Factor out core of ComputeFactorHelper
Factors out the core functionality of `ComputeFactorHelper`
(calculating the offsets of the keys in the factor state vector and the
combined state vector) into the seperate function `FactorOffsets`.

This is in support of the dense optimizer, I want to reuse this core
functionality, but don't care about the extra fluff.

Offset is intended to either be a `linearization_offsets_t` (for the
dense linearizer) or a `linearization_dense_factor_helper_t` (for the
sparse linearizer).

Topic: factor_ComputeFactorHelper
Relative: linearization_also_dense
GitOrigin-RevId: 4d3da9a7a6992127e1f560fef2cc6d5507885ace