[SymForce] Overload ComputeCovariance for MatrixX
The dense optimizer represents `hessian_lower` with a `MatrixXd`, but
`ComputeCovarianceBlockWithSchurComplement` expects a sparse matrix, and
can't be easily modified to take a dense matrix because the function
then passes it on to the `SparseSchurSolver`.

So, I add an overload that takes a dense matrix. I looked into inverting
the Schur complement, but for at least a 100x100 matrix it doesn't
actually save much time compared to calculating only the first k columns
of the full inverse. So I didn't bother with that to keep the
implementation simpler.

Topic: dense_marginalized_covariance
GitOrigin-RevId: 4be5410d320a05e477302b46ddceab00c25ef7a9