[SymForce-External] TemplateList.render supports specifying external search paths
TemplateList.render supports specifying external search paths

This change adds a `search_paths` argument such that external search paths can be specified when calling `TemplateList.render`. This allows the user to specify external directories to search when jinja resolves import/include statements in the template.

The use-case for this is to allow templates to span multiple directories - for example, some of them residing in the symforce repo and others in the user's own code.

Closes symforce-org/symforce#187

GitOrigin-RevId: 37e2d3f39679bdebb9bd337ba747c72f7f403b71