[SymForce] More stuff includes-what-it-uses
So everything I looked at in opt now compiles on its own, as far as e.g.
editor tooling is concerned.

I did remove the check we had in the sparse cholesky solver that the tcc
is only included from the header - I don't think that's a real thing to
worry about.  It'll probably work fine if someone does that, but I don't
think we really care if anyone tries.

Topic: sym-iwyu
Reviewers: bradley,chao-q,ryan-b,hayk
GitOrigin-RevId: b0f01fab28f8501cdcfa5c56f8c9149974dcf0b7