Fix symengine_wrapper search
So that it works on different platforms that e.g. use `local/lib` or
`site-packages`.

We were doing this correctly for editable installs but not non-editable
installs

Longer-term I'm leaning more and more towards requiring building
our copy of symengine separately first, but this is an easy strict
improvement for now

Fixes #330

Topic: sf-wrapper-glob
Reviewers: bradley,nathan,chao,ryan-b