[SymForce-External] Fix symengine_wrapper search
Fix symengine_wrapper search

So that it works on different platforms that e.g. use `local/lib` or
`site-packages`.

We were doing this correctly for editable installs but not non-editable
installs

Longer-term I'm leaning more and more towards requiring building
our copy of symengine separately first, but this is an easy strict
improvement for now

Fixes symforce-org/symforce#330

Topic: sf-wrapper-glob
Reviewers: bradley,nathan,chao,ryan-b
Closes symforce-org/symforce#342
GitOrigin-RevId: fce960b93a984811754da248fed658641fa16367