[SymForce] Mark test_databuffer_factor failing
This happens to pass if you use a databuffer small enough that it can be
treated as a vector, since in generated python code, both databuffers
and vectors are numpy arrays.

In reality, it doesn't work, since we don't have proper support for
DataBuffers in Values.

Related: https://github.com/symforce-org/symforce/issues/348#issuecomment-1618168635

Reviewers: bradley,nathan,chao,ryan-b,harrison,hayk
Topic: sf-databuffer-failing
GitOrigin-RevId: ce1e5cd6710873e241e3f37aab0e33acc1f14956