[SymForce] Exit if `-tol / 10 < relative_reduction < tol`
Instead of if `0 < relative_reduction < tol`.  This seems to be more
common on sim problems with no noise, but this stops the optimizer from
spending lots of iterations with very small negative relative reduction
and then exiting with failure because lambda diverges.

Ceres just uses the absolute value:
http://ceres-solver.org/nnls_solving.html#_CPPv4N5ceres6Solver7Options18function_toleranceE

But, for looser convergence thresholds a smaller value on the negative
side seems to work better?  In particular, allowing termination with a
negative relative reduction is mostly to help with cases where you're
limited by precision issues, rather than actually trying to take an
uphill step.  There isn't really an obvious function of `epsilon` to use
here instead though.

Topic: sf-negative-rel-reduction
GitOrigin-RevId: 6f0cd450264ff4a5a87e60c80b5ff44dd87fe476