[SkyMarshal] Fix printing char-like things
The LCM boolean, int8_t, uint8_t, and byte are aliases of signed or
unsigned char in C++, so they get printed as ascii 0/1/etc, which is
silly, especially since ascii 0 and 1 aren't even printable characters.
Print bools as bools, and the others as ints.  LCM doesn't have a char
type, so if you wanted that you're out of luck.

Topic: skymarshal-print-bools
Reviewers: nikhil,peter,danny
GitOrigin-RevId: a640504b22f5af0fd523f51f32799e9d1bc44198