[SymForce-External] Remove Werror from CMakeLists
Remove Werror from CMakeLists

Fixes symforce-org/symforce#369

Standard practice is to set `-Werror` in environments like CI so that it fails the build there, but not default it on for all builds, so that users have the best chance of building successfully on compilers not tested by CI that emit different warnings.

We should probably just turn off `-Wall` and `-Wextra` outside CI as well, and possibly get rid of `SYMFORCE_COMPILE_OPTIONS`, but I'm leaving that for now.

Closes symforce-org/symforce#370

GitOrigin-RevId: c9de1d9c208f58a14c6b56c29863024aad457577