skymarshal: LCM types can produce formatted json
The << operator for lcmtypes now produces pretty formatted json,
in a very similar manner to lcm_json.py.

The bulk of the logic is moved into a new header file lcm_templates.hpp
to make it easier to work with and edit. Types themselves just provide
fields() and format_field(), the actual json happens in templates.
This also makes it more flexible to do things with fields in the future.

I want to get rid of special casing bools in a follow-up PR, in
particular formatting of bool arrays is wrong in this one. This can
be fixed by using bool as the type for booleans.

Lcm structs now have their name in a "_struct_" field. Enums have their
name in an "_enum_" field. This provides an accessible name while
making it easier to distinguish structs and enums. Enums are formatted
with "name" and "value" just as in lcm_json.py, making it easier to
distinguish them from strings.

Eigen types are formatted using Eigen::IOFormat in a manner that's
reasonably pretty and still json-compliant. They're also distinguishable
from ordinary lists / multi-dim lists due to being "square".

Added tests in skymarshal to validate that the output is as expected.

GitOrigin-RevId: 0e7020b70b2548116191503bdd9c82d4e84a9e54