[SymForce] build symforce_gen with -ffast-math
This 1) should be safe and 2) should speed things up, e.g. for things
that rely on Horner's method.  Included an option to turn this off,
although I have it on by default, which I think is a good idea.

Relying on existing tests?  Not sure how confident we are flipping this.
Mostly uploading to run tests for now.

One thing is that this requires we're more careful about repeatedly
normalizing Rot3's, since it seems like that's no longer necessarily
idempotent (and I guess it was before).  That's a performance win anyway
though.

Topic: sym-fast-math
GitOrigin-RevId: 13bcdea25f335d0a6c5527b8b86f61181e4a367d