skymarshal: use real c++ bool as boolean type
This requires another new file in lcm so as to not
break existing lcm code in C.

We need to switch to using basic_string as the container because
vector has an annoying specialization for bool that uses bitset,
and is thus incompatible with out encode / decode funcs.

This makes typing cleaner and less confusing in cpp, and
also fixes lcm_json for boolean arrays.

GitOrigin-RevId: 12e30b71aaf921e17dff8dba746f8d1e847a8c83