[SymForce] IMU Preintegration symmetric covariance
Assumes the covariance is symbolically symmetric

Assumes the output is symmetric, and only looks at its lower triangle
(this saves most of the ops).  Also assumes the input covariance is
symmetric (this saves 8 ops on top of the above, and only for the auto
derivative version, the handwritten derivative version is unchanged)

Motivated by numerical concerns, but this didn't seem to make a
difference there, so it's just an efficiency gain

Also adds `Matrix.symmetric_copy`.  Thought about calling this
`self_adjoint_view` to match Eigen, but it's symmetric not self-adjoint
(not that we use any complex matrices), and it's a copy not a view, so
here we are

Topic: sf-imu-symmetric-cov
Reviewers: xipeng,karl,veeresh,dominic,peter,brad,ryan-b
GitOrigin-RevId: 97c58b2bf465adb2495d4eaaf37dc06b4d5c2b5c