[SymForce] Add dynamic lambda update
This adds an option to update lambda smoothly, using the approach from
https://www.imm.dtu.dk/documents/ftp/tr99/tr05_99.pdf.

This primarily does two things:

1) Updates lambda more smoothly when successful steps (steps that reduce
   the error) are taken.  When a step is successful, but the gain ratio
   is small, the trust region is reduced (by an amount depending on the
   gain ratio).
2) Updates lambda much more drastically when the error cannot be
   reduced, with an exponential curve in log space.  This means that a
   good starting lambda can still be found quickly, without requiring a
   large `lambda_up_factor`.

This works much better on some problems.  I don't have a good sense of
whether there are problems where this works worse.

I've added some notes in the documentation in `symforce.lcm` on this,
but for example on the BAL problem `problem-50-20431-pre.txt` this
converges in 27 iterations instead of 36 before.  On the SymForce bundle
adjustment example, however, it converges in 8 iterations both before
and after.  On the pose smoothing problem in symforce_optimizer_test, it
converges more slowly, because the default params reduce lambda to 0
more quickly close to the optimum.

Reviewers: xipeng,karl,dominic,veeresh,peter,harrison,nathan,ryan-b
Topic: sf-dynamic-lambda
GitOrigin-RevId: 25e6a6be27c6588c1156dd76b06e68208235df51