[Symforce] Add param to early exit LM optimizer if error is low
This PR adds a param to exit the LM optimization early if the error goes
below a user-specified limit. This is useful if, e.g., the optimization
is expected to possibly converge to a point where the absolute error is
near zero; in these cases the relative change in error between
iterations can appear high even though the overall error is sufficiently
low to consider the optimization complete.

By default the param is initialized to zero, which should have no effect
on existing optimizer behavior for problems which don't set the param.

Topic: symforce_lm_min_abs_error
GitOrigin-RevId: 19d30fd0e4a973edc0afc94a47d91d215ba1aef6