Allow multiple shared types with different structures
Not sure if we want to merge this as-is or rework it somehow, but I've
discovered a weird scenario I'm not sure how to deal with.

I want to codegen a function which takes in an LCM type and outputs the
same LCM type, and in the function copies only certain fields. The LCM
types are quite large, so writing this code by hand is tedious. I've
done this by codegening a function where the output Values is the same
as the input Values but with the fields I don't want to copy deleted.
Both the input and output Values are then included in the `shared_types`
dict, and both point to the same LCM type.

However, during codegen symforce asserts that the index of a
input/output Values must match the index of any previously seen
input/output Values of the same type. Because I used the same shared
type for my input and output Values, this assert fails because they each
have a different index (specifically, the output Values is missing some
elements).

This PR is basically a work-around which ignores these assertions for
shared types. In some cases this could allow users to generate code with
can't compile, but this is true whenever specifying shared types anyway
(since there's no guarantee the shared type actually matches the
codegened code), so I think this is fine. Basically "if you're using
shared types, you probably know what you're doing so the assertion
safe-guards are removed".

Topic: allow_diff_shared_type_index
GitOrigin-RevId: 6b08c82eb4ca85b29e275f1bcc783a2c7b71c3e9